Enable vulkan on the Raspberry Pi 5
This is necessary for smooth video decoding on the Raspberry Pi 5.

The patch also adds a FORCE_VULKAN env variable, which can be used
to override the default behavior and force vulkan (not) to be loaded.
This is useful because enabling vulkan on the Raspberry Pi 5 comes
with a 20% performance penalty to OpenGL, which might not be a good
tradeoff for all users.

Change-type: patch
Signed-off-by: Michal Toman <michalt@balena.io>